src/bookableu-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic[email]
python-dotenv
boto3
passlib[bcrypt]
pyjwt
PyMuPDF>=1.19.0
ebooklib
pdfminer.six
numpy
scikit-learn
faiss-cpu
python-multipart
openai
pytest
python-jose